% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrValues.R
\name{mlrValues}
\alias{mlrValues}
\title{Multiple linear regression value calculator}
\usage{
mlrValues(formula, df)
}
\arguments{
\item{formula}{String in the format of: "y~x"}

\item{df, }{Dataframe}
}
\description{
Generates a multiple linear model and outputs associated values and statistics.
}
\examples{
mlrValues()
}
\keyword{linear}
\keyword{mlr,}
\keyword{multiple}
\keyword{of}
\keyword{regression,}
\keyword{rss,}
\keyword{squares,}
\keyword{sse,}
\keyword{ssreg,}
\keyword{sum}
\keyword{sxx}
