% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slrValues.R
\name{slrValues}
\alias{slrValues}
\title{Simple linear regression value calculator}
\usage{
slrValues(y, x, df)
}
\arguments{
\item{y}{String response variable}

\item{x}{String explanatory variable}

\item{df}{Dataframe}
}
\description{
Generates a  simple linear model and outputs associated values and statistics.
}
\examples{
slrValues()
}
\keyword{linear}
\keyword{of}
\keyword{regression,}
\keyword{rss,}
\keyword{simple}
\keyword{slr,}
\keyword{squares,}
\keyword{sse,}
\keyword{ssreg,}
\keyword{sum}
\keyword{sxx}
